refactor(sign): reorganiza templates e ajusta exibição de produtos
  - Move os templates para subpastas por recurso (products/ e
  manufacturers/) e atualiza os template_name das views
  - Troca quantity de DecimalField para PositiveIntegerField (+ migration)
  - Exibe a sigla do tipo de unidade (unit_type) no lugar do label
  completo na listagem, detalhe e no select do formulário
  - Destaca nome, fabricante, código de barras e código do fabricante
  na tela de exclusão de produto
  - Alinha os botões dos forms de edição/exclusão à direita
  - Atualiza docs/produtos.md

+++ sign/models.py
@@ -46,9 +46,7 @@ class Product(models.Model):
     manufacturer_code = models.CharField(
         "Código do fabricante", max_length=64, blank=True
     )
-    quantity = models.DecimalField(
-        "Quantidade", max_digits=12, decimal_places=3, default=0
-    )
+    quantity = models.PositiveIntegerField("Quantidade", default=0)
     unit_type = models.CharField(
         "Tipo de unidade",
         max_length=4,
